Split editor and re-organize components
